requirements.txt
django>=5.0
django-ninja>=1.1
django-cors-headers
python-dotenv
crewai
langchain
pydantic>=2.0
pymongo
uvicorn
gunicorn
requests
openai
PyYAML
beautifulsoup4
